Added event queue to config, on config saved send new initial info if protobuf protocol is enabled.
